requirements.txt
alabaster==0.7.11
autopep8==1.3.5
Babel==2.6.0
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
docutils==0.14
feedparser==5.2.1
Flask==1.0.2
Flask-Compress==1.4.0
Flask-HTMLmin==1.3.2
Flask-PyMongo==2.0.0
google-api-python-client==1.7.4
google-auth==1.5.0
google-auth-httplib2==0.0.3
htmlmin==0.1.12
httplib2==0.11.3
idna==2.7
imagesize==1.0.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
oauth2client==4.1.2
packaging==17.1
pprint==0.1
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pydocstyle==2.1.1
pyflakes==2.0.0
Pygments==2.2.0
pylama==7.4.3
pymongo==3.7.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-twitch-client==0.5.1
pytz==2018.5
PyYAML==3.13
requests==2.19.1
rsa==3.4.2
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.6
sphinxcontrib-websupport==1.1.0
uritemplate==3.0.0
urllib3==1.23
Werkzeug==0.14.1
